+++ analysis/filter/MergeFilter/src/test/kotlin/de/maibornwolff/codecharta/filter/mergefilter/MergeFilterTest.kt
@@ -1,20 +1,16 @@
 package de.maibornwolff.codecharta.filter.mergefilter
 
 import de.maibornwolff.codecharta.filter.mergefilter.MergeFilter.Companion.main
-import de.maibornwolff.codecharta.model.Project
-import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.util.InputHelper
 import io.mockk.every
 import io.mockk.mockkObject
 import io.mockk.unmockkAll
-import io.mockk.verify
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import picocli.CommandLine
 import java.io.ByteArrayOutputStream
 import java.io.File
-import java.io.FileInputStream
 import java.io.PrintStream
 
 class MergeFilterTest {
@@ -98,26 +94,18 @@ class MergeFilterTest {
     }
 
     @Test
-    fun `should abort if at least one input file does not exist`() {
+    fun `should not execute merge if input is invalid`() {
         mockkObject(InputHelper)
         every {
-            InputHelper.getAndCheckAllSpecifiedInputFiles(any())
-        } returns mutableListOf()
-
-        mockkObject(ProjectDeserializer)
-        every {
-            ProjectDeserializer.deserializeProject(any<FileInputStream>())
-        } returns Project("")
+            InputHelper.isInputValid(any(), any())
+        } returns false
 
         System.setErr(PrintStream(errContent))
         CommandLine(MergeFilter()).execute(
                 "src/test/resources/mergeFolderTest/file1.cc.json",
-                "src/test/resources/mergeFolderTest/file2.cc.json",
                 "src/test/resources/thisDoesNotExist.cc.json").toString()
         System.setErr(originalErr)
 
-        assertThat(errContent.toString()).contains("Aborting execution because one or more input files have not been found or no input was specified at all!")
-
-        verify(exactly = 0) { ProjectDeserializer.deserializeProject(any<FileInputStream>()) }
+        assertThat(errContent.toString()).contains("Input invalid files/folders for MergeFilter, stopping execution")
     }
 }