+++ analysis/import/CSVImporter/src/main/kotlin/de/maibornwolff/codecharta/importer/csv/CSVImporter.kt
@@ -4,6 +4,7 @@ import de.maibornwolff.codecharta.serialization.ProjectSerializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
 import de.maibornwolff.codecharta.tools.interactiveparser.util.InteractiveParserHelper
+import de.maibornwolff.codecharta.util.InputHelper
 import picocli.CommandLine
 import java.io.File
 import java.io.IOException
@@ -43,6 +44,9 @@ class CSVImporter(
 
     @Throws(IOException::class)
     override fun call(): Void? {
+        if (!InputHelper.isInputValid(files.toTypedArray(), canInputContainFolders = false)) {
+            throw IllegalArgumentException("Input invalid file for CSVImporter, stopping execution...")
+        }
 
         val csvProjectBuilder = CSVProjectBuilder(pathSeparator, csvDelimiter, pathColumnName)
         files.map { it.inputStream() }.forEach<InputStream> { csvProjectBuilder.parseCSVStream(it) }
@@ -54,6 +58,7 @@ class CSVImporter(
     }
 
     companion object {
+
         @JvmStatic
         fun main(args: Array<String>) {
             CommandLine(CSVImporter()).execute(*args)